refactor: move EPUB translation into pipeline
