+++ pdf_craft/extractor/ocr/__init__.py
@@ -1,4 +1,1 @@
-from pdf_craft.pdf.ocr import OCR, OCREvent, OCREventKind
-from pdf_craft.ocr_config import *  # noqa: F403
-
-__all__ = ["OCR", "OCREvent", "OCREventKind"]
+"""OCR source adapter namespace."""
