+++ pdf_craft/extractor/pdf/__init__.py
@@ -1,1 +1,1 @@
-from pdf_craft.pdf import *  # noqa: F403
+"""PDF source adapter namespace."""
