+++ pdf_craft/extractor/toc/__init__.py
@@ -1,1 +1,1 @@
-from pdf_craft.toc import *  # noqa: F403
+"""TOC analysis namespace."""
