+++ pdf_craft/renderer/__init__.py
@@ -1,4 +1,1 @@
-from .epub import render_epub_file
-from .markdown import render_markdown_file
-
-__all__ = ["render_epub_file", "render_markdown_file"]
+"""Document rendering boundary for Markdown and EPUB targets."""
