+++ pdf_craft/renderer/epub/__init__.py
@@ -1,3 +1,1 @@
-from pdf_craft.epub.render import render_epub_file
-
-__all__ = ["render_epub_file"]
+"""EPUB renderer namespace."""
