+++ pdf_craft/renderer/markdown/__init__.py
@@ -1,3 +1,1 @@
-from pdf_craft.markdown.render.render import render_markdown_file
-
-__all__ = ["render_markdown_file"]
+"""Markdown renderer namespace."""
